Moved to stdint.h types

+++ src/irq.c
@@ -27,13 +27,13 @@ void *irq_routines[16] =
 };
 
 /* This installs a custom IRQ handler for the given IRQ */
-void irq_install_handler(int irq, void (*handler)(struct regs *r))
+void irq_install_handler(size_t irq, void (*handler)(struct regs *r))
 {
     irq_routines[irq] = handler;
 }
 
 /* This clears the handler for a given IRQ */
-void irq_uninstall_handler(int irq)
+void irq_uninstall_handler(size_t irq)
 {
     irq_routines[irq] = 0;
 }
@@ -67,22 +67,22 @@ void irq_install()
 {
     irq_remap();
 
-	idt_set_gate(32, (unsigned)irq0, 0x08, 0x8E);
-	idt_set_gate(33, (unsigned)irq1, 0x08, 0x8E);
-	idt_set_gate(34, (unsigned)irq2, 0x08, 0x8E);
-	idt_set_gate(35, (unsigned)irq3, 0x08, 0x8E);
-	idt_set_gate(36, (unsigned)irq4, 0x08, 0x8E);
-	idt_set_gate(37, (unsigned)irq5, 0x08, 0x8E);
-	idt_set_gate(38, (unsigned)irq6, 0x08, 0x8E);
-	idt_set_gate(39, (unsigned)irq7, 0x08, 0x8E);
-	idt_set_gate(40, (unsigned)irq8, 0x08, 0x8E);
-	idt_set_gate(41, (unsigned)irq9, 0x08, 0x8E);
-	idt_set_gate(42, (unsigned)irq10, 0x08, 0x8E);
-	idt_set_gate(43, (unsigned)irq11, 0x08, 0x8E);
-	idt_set_gate(44, (unsigned)irq12, 0x08, 0x8E);
-	idt_set_gate(45, (unsigned)irq13, 0x08, 0x8E);
-	idt_set_gate(46, (unsigned)irq14, 0x08, 0x8E);
-	idt_set_gate(47, (unsigned)irq15, 0x08, 0x8E);
+	idt_set_gate(32, (uint32_t)irq0, 0x08, 0x8E);
+	idt_set_gate(33, (uint32_t)irq1, 0x08, 0x8E);
+	idt_set_gate(34, (uint32_t)irq2, 0x08, 0x8E);
+	idt_set_gate(35, (uint32_t)irq3, 0x08, 0x8E);
+	idt_set_gate(36, (uint32_t)irq4, 0x08, 0x8E);
+	idt_set_gate(37, (uint32_t)irq5, 0x08, 0x8E);
+	idt_set_gate(38, (uint32_t)irq6, 0x08, 0x8E);
+	idt_set_gate(39, (uint32_t)irq7, 0x08, 0x8E);
+	idt_set_gate(40, (uint32_t)irq8, 0x08, 0x8E);
+	idt_set_gate(41, (uint32_t)irq9, 0x08, 0x8E);
+	idt_set_gate(42, (uint32_t)irq10, 0x08, 0x8E);
+	idt_set_gate(43, (uint32_t)irq11, 0x08, 0x8E);
+	idt_set_gate(44, (uint32_t)irq12, 0x08, 0x8E);
+	idt_set_gate(45, (uint32_t)irq13, 0x08, 0x8E);
+	idt_set_gate(46, (uint32_t)irq14, 0x08, 0x8E);
+	idt_set_gate(47, (uint32_t)irq15, 0x08, 0x8E);
 }
 
 /* Each of the IRQ ISRs point to this function, rather than
@@ -120,3 +120,4 @@ void irq_handler(struct regs *r)
     *  interrupt controller too */
     outportb(0x20, 0x20);
 }
+
